MOS-23729: Message sender changes
Former-commit-id: a34b6d5bc58c37e9599ca8cea300717620cf38c4
Former-commit-id: b728a408753a627d35fe840609085ed359ad60b7

+++ registration-processor/post-processor/registration-processor-printing-stage/src/main/java/io/mosip/registration/processor/print/stage/PrintStage.java
@@ -230,8 +230,8 @@ public class PrintStage extends MosipVerticleAPIManager {
 			registrationStatusDto.setRegistrationStageName(this.getClass().getSimpleName());
 
 			JSONObject jsonObject = utilities.retrieveUIN(regId);
-			Integer value=JsonUtil.getJSONValue(jsonObject, UIN);
-			String uin= Integer.toString(value);
+			Long value=JsonUtil.getJSONValue(jsonObject, UIN);
+			String uin= Long.toString(value);
 			
 			
 			Map<String, byte[]> documentBytesMap = printService.getDocuments(IdType.RID, regId);
