+++ registration-processor/registration-processor-print-service-impl/src/main/java/io/mosip/registration/processor/print/service/impl/PrintServiceImpl.java
@@ -188,8 +188,8 @@ public class PrintServiceImpl implements PrintService<Map<String, byte[]>> {
 				uin = idValue;
 			} else if (idType.toString().equalsIgnoreCase(RID)) {
 				JSONObject jsonObject = utilities.retrieveUIN(idValue);
-				Integer value=JsonUtil.getJSONValue(jsonObject, UIN);
-				uin= Integer.toString(value);
+				Long value=JsonUtil.getJSONValue(jsonObject, UIN);
+				uin= Long.toString(value);
 				if (uin == null) {
 					regProcLogger.error(LoggerFileConstant.SESSIONID.toString(),
 							LoggerFileConstant.REGISTRATIONID.toString(), null,
